CCCP-345, chore: cleanup codebase (#112)
* CCCP-345, chore: cleanup codebase

* CCCP-345, deps: update substrate crates

* CCCP-345, fix: remove v2 feature

* CCCP-345, test: add socket msg decode test

* CCCP-345, fix: check if pending tx exists

* CCCP-345, fix: case of `couldn't even find tx in txpool`

* CCCP-345, fix: remove unused code and add missing doc

* CCCP-345, refactor: cleanup file paths

* CCCP-345, refactor: rename periodic

* CCCP-345, chore: set eip1559 as default for bsc

* CCCP-345, deps: release v1.1.1

* CCCP-345, chore: apply clippy

* CCCP-345, chore: pub update

* CCCP-345, chore: unify asset oracle id format

* CCCP-345, deploy: builder rust image to latest

---------
