+++ client/Cargo.toml
@@ -20,9 +20,9 @@ async-trait = { workspace = true }
 async-recursion = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 tokio = { workspace = true, features = [
-    "signal",
-    "rt-multi-thread",
-    "parking_lot",
+	"signal",
+	"rt-multi-thread",
+	"parking_lot",
 ] }
 k256 = { workspace = true, features = ["ecdsa"] }
 sha3 = { workspace = true }
@@ -36,6 +36,3 @@ sc-service = { workspace = true }
 
 [dev-dependencies]
 serde_json = { workspace = true }
-
-[features]
-v2 = []
