+++ client/src/eth/wallet.rs
@@ -1,4 +1,5 @@
-use br_primitives::{errors::INVALID_PRIVATE_KEY, eth::ChainID};
+use br_primitives::{constants::errors::INVALID_PRIVATE_KEY, eth::ChainID};
+
 use ethers::{
 	prelude::k256::ecdsa::SigningKey,
 	signers::{LocalWallet, Signer},
@@ -15,8 +16,12 @@ use sha3::{Digest, Keccak256};
 type WalletResult<T = ()> = Result<T, Box<dyn std::error::Error + Send + Sync>>;
 
 #[derive(Debug)]
+/// The component that contains anEthereum private-public key pair
+/// which can be used for signing messages.
 pub struct WalletManager {
+	/// The wallet instantiated with a locally stored private key.
 	pub signer: ethers::signers::Wallet<SigningKey>,
+	/// The ECDSA/secp256k1 signing key.
 	secret_key: Option<K256SigningKey>,
 }
