+++ configs/config.testnet.yaml
@@ -28,7 +28,7 @@ evm_providers:
     call_interval: 3000
     block_confirmations: 5
     is_relay_target: false
-    is_initially_escalated: true
+    eip1559: true
     socket_address: "0x8039c3AD8ED55509fD3f6Daa78867923fDe6E61c"
     authority_address: "0xCf9f6428A309b6652a1dfaA4d8aB8B61C9c7E8CF"
     chainlink_usdc_usd_address: "0x90c069C4538adAc136E051052E14c1cD799C41B7"
