+++ metrics/Cargo.toml
@@ -14,8 +14,5 @@ lazy_static = { workspace = true }
 sentry = { workspace = true }
 ethers = { workspace = true }
 
-# Bifrost Relayer
-br-primitives = { path = "../primitives", default-features = false }
-
 # Substrate
 prometheus-endpoint = { workspace = true }
