+++ periodic/src/lib.rs
@@ -1,7 +1,11 @@
-pub mod heartbeat_sender;
-pub mod price_feeder;
+mod heartbeat_sender;
+mod price_feeder;
 mod price_source;
-pub mod roundup_emitter;
-pub mod socket_rollback_handler;
+mod roundup_emitter;
+mod socket_rollback_emitter;
+pub mod traits;
 
+pub use heartbeat_sender::*;
 pub use price_feeder::*;
+pub use roundup_emitter::*;
+pub use socket_rollback_emitter::*;
