+++ periodic/src/price_source/coingecko.rs
@@ -5,12 +5,9 @@ use reqwest::{Response, Url};
 use serde::Deserialize;
 use tokio::time::{sleep, Duration};
 
-use br_primitives::{
-	periodic::{PriceFetcher, PriceResponse},
-	sub_display_format,
-};
+use br_primitives::{periodic::PriceResponse, sub_display_format};
 
-use crate::price_source::LOG_TARGET;
+use crate::{price_source::LOG_TARGET, traits::PriceFetcher};
 
 const SUB_LOG_TARGET: &str = "coingecko";
