+++ primitives/src/contracts/mod.rs
@@ -2,4 +2,3 @@ pub mod authority;
 pub mod chainlink_aggregator;
 pub mod relayer_manager;
 pub mod socket;
-pub mod vault;
