+++ primitives/src/contracts/socket.rs
@@ -142,17 +142,17 @@ pub fn get_asset_oids() -> BTreeMap<&'static str, H256> {
 		),
 		(
 			"DAI",
-			H256::from_str("0x010001000000000000000000000000000000000000000000000000000000000b")
+			H256::from_str("010001000000000000000000000000000000000000000000000000000000000b")
 				.unwrap(),
 		),
 		(
 			"BTCB",
-			H256::from_str("0x010001000000000000000000000000000000000000000000000000000000000c")
+			H256::from_str("010001000000000000000000000000000000000000000000000000000000000c")
 				.unwrap(),
 		),
 		(
 			"WBTC",
-			H256::from_str("0x010001000000000000000000000000000000000000000000000000000000000d")
+			H256::from_str("010001000000000000000000000000000000000000000000000000000000000d")
 				.unwrap(),
 		),
 	])
