--- primitives/src/contracts/vault.rs
@@ -1,7 +1,0 @@
-use ethers::prelude::abigen;
-
-abigen!(
-	VaultContract,
-	"../abi/abi.vault.json",
-	event_derives(serde::Deserialize, serde::Serialize)
-);
