+++ primitives/src/lib.rs
@@ -2,13 +2,9 @@ pub mod bootstrap;
 pub mod cli;
 pub mod constants;
 pub mod contracts;
-pub mod errors;
 pub mod eth;
 pub mod periodic;
-
-pub use contracts::*;
-pub use errors::*;
-pub use periodic::*;
+pub mod tx;
 
 pub fn sub_display_format(log_target: &str) -> String {
 	format!("{:<019}", log_target)
