+++ relayer/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bifrost-relayer"
-version = "1.1.0"
+version = "1.1.1"
 description = "The core rust implementation for Bifrost relayer"
 authors = { workspace = true }
 homepage = { workspace = true }
