+++ relayer/src/verification.rs
@@ -1,12 +1,14 @@
 use br_primitives::{
 	cli::Configuration,
 	constants::{
-		DEFAULT_BOOTSTRAP_ROUND_OFFSET, DEFAULT_DUPLICATE_CONFIRM_DELAY_MS,
-		DEFAULT_ESCALATE_PERCENTAGE, DEFAULT_GET_LOGS_BATCH_SIZE, MAX_BLOCK_CONFIRMATIONS,
-		MAX_BOOTSTRAP_ROUND_OFFSET, MAX_CALL_INTERVAL_MS, MAX_DUPLICATE_CONFIRM_DELAY_MS,
-		MAX_ESCALATE_PERCENTAGE, MIN_GET_LOGS_BATCH_SIZE,
+		cli::{
+			DEFAULT_BOOTSTRAP_ROUND_OFFSET, DEFAULT_DUPLICATE_CONFIRM_DELAY_MS,
+			DEFAULT_ESCALATE_PERCENTAGE, DEFAULT_GET_LOGS_BATCH_SIZE, MAX_BLOCK_CONFIRMATIONS,
+			MAX_BOOTSTRAP_ROUND_OFFSET, MAX_CALL_INTERVAL_MS, MAX_DUPLICATE_CONFIRM_DELAY_MS,
+			MAX_ESCALATE_PERCENTAGE, MIN_GET_LOGS_BATCH_SIZE,
+		},
+		errors::PARAMETER_OUT_OF_RANGE,
 	},
-	PARAMETER_OUT_OF_RANGE,
 };
 
 /// Verifies whether the certain numeric parameters specified in the configuration YAML file are valid.
